+++ Sources/StreamChat/APIClient/Endpoints/EndpointPath+OfflineRequest.swift
@@ -25,6 +25,7 @@ extension EndpointPath {
              .channels,
              .channelUpdate,
              .connect,
+             .custom,
              .createCall,
              .createChannel,
              .createDevice,